// apps/backend/src/core/application/search/enums/filter-operator.enum.ts
export enum FilterOperator {
  EQUALS = 'equals',

  NOT_EQUALS = 'notEquals',

  CONTAINS = 'contains',

  STARTS_WITH = 'startsWith',

  ENDS_WITH = 'endsWith',

  GREATER_THAN = 'greaterThan',

  GREATER_THAN_OR_EQUAL = 'greaterThanOrEqual',

  LESS_THAN = 'lessThan',

  LESS_THAN_OR_EQUAL = 'lessThanOrEqual',

  IN = 'in',

  IS_NULL = 'isNull',

  IS_NOT_NULL = 'isNotNull',
}